backend/requirements.txt
alembic==1.11.1
anyio==4.2.0
asyncpg==0.28.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
fastapi==0.100.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pyasn1==0.5.1
pydantic==1.10.13
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a37
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.22.0
